Fix compilation warning.

+++ src/cpp/microstrain/array_view.hpp
@@ -64,7 +64,7 @@ struct ArrayView
     constexpr element_type back() const noexcept { return *m_ptr[extent-1]; }
 
     template<typename Index>  // Required because otherwise operator T* would make this ambiguous
-    constexpr reference operator[](Index idx) const noexcept { assert(idx < extent); return m_ptr[idx]; }
+    constexpr reference operator[](Index idx) const noexcept { assert(static_cast<size_t>(idx) < extent); return m_ptr[idx]; }
 
     constexpr pointer data() const noexcept { return m_ptr; }
 
@@ -116,7 +116,7 @@ struct ArrayView<T, DYNAMIC_EXTENT>
     constexpr element_type back() const noexcept { return *m_ptr[m_cnt-1]; }
 
     template<typename Index>  // Required because otherwise operator T* would make this ambiguous
-    constexpr reference operator[](Index idx) const noexcept { assert(idx < m_cnt); return m_ptr[idx]; }
+    constexpr reference operator[](Index idx) const noexcept { assert(static_cast<size_t>(idx) < m_cnt); return m_ptr[idx]; }
 
     constexpr pointer data() const noexcept { return m_ptr; }
